Added more context to functions
For the addListItem() function I defined a variable that would be the inserted text for the Span element. Then I appended the entire content within a single li

For the checklist item; I figured that all we are changing is the class from shopping-item_checked or back to shopping-item depending on what the current status is.

For the removeList Item added I am playing around with event delegation to see how we would be able to remove all listed contents that are children under the Li element.